+++ src/app/page.js
@@ -1,6 +1,6 @@
 "use client";
 import AddProductForm from "@/components/AddProducts/AddProductForm";
-import style from "./page.module.css";
+import style from "@/css/AddProducts/page.module.css";
 
 export default function Home() {
   return (
