+++ src/components/AddProducts/ProductComponents/DeleteButton.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { AiOutlineMinus } from "react-icons/ai";
-import "@/app/AddProductForm.css";
+import "@/css/AddProducts/AddProductForm.css";
 
 const DeleteButton = ({ index, onClick }) => {
   return (
